ex8_ch_3.py
Instructions: Write a program to evenly divide pizzas. Prompt for the number of people, the number of pizzas, and the number of slices per pizza. Ensure that the number of pieces comes out even. Display the number of pieces of pizza each person should get. If there are leftovers, show the number of leftover pieces.

Example output: 
How many people? 8
       How many pizzas do you have? 2
       8 people with 2 pizzas
       Each person gets 2 pieces of pizza.
       There are 0 leftover pieces.