ex_9_ch_3.py
Instructions: Calculate gallons of paint needed to paint the ceiling of a room. Prompt for the length and width, and assume one gallon covers 350 square feet. Display the number of gallons needed to paint the ceiling as a whole number.

Example output: You will need to purchase 2 gallons of paint to cover 360 square feet.