capstone_lesson6_v1.py
Instructions: convert hours to minutes and print hours and remaining minutes.

The capstone project gave me the total number of minutes, so I created two programs (capstone_lesson6_v1.py and capstone_lesson6_v2.py) starting with the total number of minutes and solved for the total hours and the total remaining minutes. 

Then, for fun, I created a program that asks the user for the total number of minutes (capstone_lesson6_v3.py). I convert the total minutes to hours and remaining minutes, and print both.