ex1_ch1.py
Instructions: Create a simple tip calculator. The program should prompt for a bill amount and a tip rate. The program must compute the tip and then display both the tip and the total amount of the bill.

Example output:
What is the bill? $200
What is the tip percentage? 15
The tip is $30.00
The total is $230.00