ex10_ch3.py
Instructions: 
Create a simple self-checkout system. Prompt for the prices and quantities of three items. Calculate the subtotal of the items. Then calculate the tax using a tax rate of 5.5%. Print out the line items with the quantity and total, and then print out the subtotal, tax amount, and total.

Example Output:
       Enter the price of item 1: 25
       Enter the quantity of item 1: 2
       Enter the price of item 2: 10
       Enter the quantity of item 2: 1
       Enter the price of item 3: 4
       Enter the quantity of item 3: 1
       Subtotal: $64.00
       Tax: $3.52
       Total: $67.52