ex_14_ch4.py
Instructions: Write a simple program to compute the tax on an order amount. The program should prompt for the order amount and the state. If the state is “WI,” then the order must be charged 5.5% tax. The program should display the subtotal, tax, and total for Wisconsin residents but display just the total for non-residents.

Constraints: 

• Implement this program using only a simple if state- ment—don’t use an else clause.
• Ensurethatallmoneyisroundeduptothenearestcent.
• Useasingleoutputstatementattheendoftheprogram
to display the program results.