ex15_ch4.py
Instructions: Create a simple program that validates user login credentials. The program must prompt the user for a username and password. The program should compare the password given by the user to a known password. If the password matches, the program should display “Welcome!” If it doesn’t match, the program should display “I don’t know you.”

Constraints:
• Use an if/else statement for this problem. • Make sure the program is case sensitive.