ex17_ch4_v2.py
Create a program that prompts for your weight, gender, number of drinks, the amount of alcohol by volume of the drinks consumed, and the amount of time since your last drink. Calculate your blood alcohol content (BAC) using this formula
BAC = (A × 5.14 / W × r) − .015 × H

Display whether or not it’s legal to drive by comparing the blood alcohol content to 0.08.