ex_24.py
Create a program that compares two strings and determines if the two strings are anagrams. The program should prompt for both input strings and display the output as shown in the example that follows:
Enter two strings and I'll tell you if they are anagrams:
       Enter the first string: note
       Enter the second string: tone
       "note" and "tone" are anagrams.

Constraints:
Implement the program using a function called isAna- gram, which takes in two words as its arguments and returns true or false. Invoke this function from your main program.