ex_26.py
Write a program that will help you determine how many months it will take to pay off a credit card balance. The program should ask the user to enter the balance of a credit card and the APR of the card. The program should then return the number of months needed.
The formula for this is
1 log(1 + pb(1 − (1 + i)30))
 n=− × 30
log(1 + i)
  where
• n is the number of months.
• i is the daily rate (APR divided by 365). • b is the balance.
• p is the monthly payment.